+++ routes/rentals.js
@@ -44,32 +44,38 @@ router.post("/", async (req, res) => {
     },
   });
 
-    rental.save();
-    movie.numberInStock--;
-    movie.save();
+  rental.save();
+  movie.numberInStock--;
+  movie.save();
 
-    // handling Transaction with Fawn // But Failed
+  // handling Transaction with Fawn // But Failed
 
-//   try {
-//     new Fawn.Task()
-//       .save("rentals", rental)
-//       .update(
-//         "movies",
-//         { _id: movie._id },
-//         {
-//           $inc: { numberInStock: -1 },
-//         }
-//       )
-//       .run();
-//   } catch (error) {
-//     res.status(500).send("Something went wrong");
-//   }
+  //   try {
+  //     new Fawn.Task()
+  //       .save("rentals", rental)
+  //       .update(
+  //         "movies",
+  //         { _id: movie._id },
+  //         {
+  //           $inc: { numberInStock: -1 },
+  //         }
+  //       )
+  //       .run();
+  //   } catch (error) {
+  //     res.status(500).send("Something went wrong");
+  //   }
 
   res.status(200).send(rental);
 });
 
 //Read Particular
 router.get("/:id", async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id))
+    return res
+      .status(404)
+      .send(
+        `"rentalId" with value "${req.params.id}" fails to match the valid mongo id pattern`
+      );
   const rental = await Rental.findById(req.params.id);
 
   console.log("Rental: " + rental);
@@ -84,6 +90,12 @@ router.get("/:id", async (req, res) => {
 
 // Delete
 router.delete("/:id", async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id))
+    return res
+      .status(404)
+      .send(
+        `"rentalId" with value "${req.params.id}" fails to match the valid mongo id pattern`
+      );
   const rental = await Rental.findByIdAndRemove(req.params.id);
 
   // const movie = movies.find((c) => c.id == req.params.id);
@@ -103,10 +115,16 @@ router.delete("/:id", async (req, res) => {
 //Update
 
 router.put("/:id", async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id))
+    return res
+      .status(404)
+      .send(
+        `"rentalId" with value "${req.params.id}" fails to match the valid mongo id pattern`
+      );
   return res
     .status(404)
     .send(
-      "We are not supporting Update Rental Data from Outside. Update dont make sense now 09 May 2022 ; "
+      "hahaha .!. \n We are not supporting Update Rental Data from Outside. Update API for rentals dont make sense now 09 May 2022 ; "
     );
 
   console.log("Req", req.body);
